+++ org.beangle.ems.system/src/main/java/org/beangle/ems/system/web/action/PropertyAction.java
@@ -83,7 +83,7 @@ public class PropertyAction extends ActionSupport {
   }
 
   public String remove() {
-    PropertyConfigItemBean config = entityDao.get(PropertyConfigItemBean.class, getLong("config.id"));
+    PropertyConfigItemBean config = entityDao.get(PropertyConfigItemBean.class, getInteger("config.id"));
     if (null != config) entityDao.remove(config);
     return redirect("dynaInfo", "info.save.success");
   }
